backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=65.0,<66.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
dj-database-url>=2.2,<3.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
